A relative error of 0.005 is still insufficiently precise, and can lead to a slippage relative to a quote amount shown to the user of greater than 1%.
Suppose for example that a quote is for 0.9951 BTC. That will be rounded
to 1.00 BTC (it is within 0.005 relative tolerance and at least 2 decimal
places). If it slips by 0.9% relative to the exact value, then the user
gets 0.9861441 BTC, and the slippage relative to what the user was shown
is ~1.39%, which breaks the stated contract with the user.

This changes the relative error threshold to 0.0005, effectively adding
up to one more decimal place if necessary. The slippage can still be up
to 1.05% but we could fix that on the NEAR side by requiring the actual
slippage to be < 0.95% (whereas it is not reasonable to require the
actual slippage to be < 0.5% just to account for rounding).

We also fix two additional subtle bugs:
* the scale of the returned value could be less than 8 decimal places,
  potentially leading to undesired precision loss in subsequent arithmetic;
* if the original value had a scale > 8 decimal places and it was not
  possible to meet the error threshold with 8 decimal places, then the
  error threshold could have been violated.

The changed rounding is applied everywhere but that's fine.

Signed-off-by: Daira-Emma Hopwood <daira@jacaranda.org>