Upgrade Android Gradle plugin to 8.13.0
This fixed a dependency conflict in the Macrobenchmark tests. The runner
underneath it depends on gRPC, in particular the internal class
`io.grpc.InternalGlobalInterceptors`. This class was removed during a
refactor in grpc/grpc-java#11148, released in `grpc-api 1.65.0`.
However, due to the Zcash Android SDK depending on gRPC itself, we ended
up with a version conflict: `grpc-core 1.68.2` and `grpc-api 1.57.2`
were the resolved versions, and due to the aforementioned refactor they
were incompatible.

After the upgrade to Android Gradle plugin 8.13.0, we consistently use
gRPC 1.69.1, which avoids any version conflicts.